+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix PollResult voted bug
 - Move GET request bodies to query for web
 - Fix poll DateUpdated set to time zero value on creation
+- Use group polls notification topic
 
 ## [1.10.0] - 2025-05-15
 ### Fixed